arm: sun4i: switch to PLL-based div choice
On boot, we get handed a frequency with weird divs (1008Mhz, 1-2-2-8).
As 1-2-2-8 is not on the tables, we pick the last entry incorrectly,
and this causes issues. A similar issue seems to happen on suspend.

By choosing the divs based on the PLL value, we can handle all these
cases gracefully